refactor(frontend): extract PresetFormBody.vue — eliminates 200+ LOC duplication (#9060)
Extracted shared preset form implementation from PresetsSettingsPanel and PresetManagementModal into PresetFormBody.vue component.

Before:
- PresetsSettingsPanel.vue: 447 lines (script + template + styles duplicated)
- PresetManagementModal.vue: 457 lines (script + template + styles duplicated)

After:
- PresetFormBody.vue: 307 lines (extracted shared component)
- PresetsSettingsPanel.vue: 8 lines (uses PresetFormBody)
- PresetManagementModal.vue: 24 lines (uses PresetFormBody + BaseModal wrapper)

Net reduction: ~594 lines → ~339 lines (43% reduction)

Duplication eliminated:
- Reactive state management (form, errors, loading states)
- All methods (resetForm, startEdit, validate, savePreset, confirmDelete, executeDelete)
- Complete form template (3 fields + actions)
- Preset list template + loading/empty states
- Delete confirmation dialog
- All scoped styles
